backend/requirements.txt
fastapi
uvicorn
pymongo
motor
python-dotenv
bcrypt
python-jose[cryptography]
pydantic
biopython
scikit-learn
torch
transformers
passlib
python-multipart

matplotlib
numpy
pandas
requests
